fix api error handling
Since the switch to go-service in 33fbc4506889aa9f088f791374997e0dcf4d5183, we
use shttp.JSONError for API errors. But this new structure use "message" for
the message field instead of "error" as we used in Eventline before.

So we switch to our own type and HTTP error handler to maintain backward
compatiblity.